separate tests using mocks from other tests in fortnite_api
